feat(v2): support --fix-script on rescue for fix-then-inspect
Move the startup-script composer to orchestration/compose.py so both
orchestrators can use it without a circular import (repair re-exports for
backward compatibility). When rescue gets --fix-script, the orchestrator
appends the custom script after the mount part with the completion marker
relocated to the end, then leaves the VM in rescue mode for inspection.
Windows VMs are blocked at pre-flight (Linux only for now); the
confirmation lists the script that will run.
